+++ src/OperationalInsights/OperationalInsights.Test/ScenarioTests/Common.ps1
@@ -63,3 +63,14 @@ function Get-ProviderLocation()
 
 	return "East US"
 }
+
+<#
+.SYNOPSIS
+Cleans the created resource groups
+#>
+function Clean-ResourceGroup($rgname)
+{
+    if ([Microsoft.Azure.Test.HttpRecorder.HttpMockServer]::Mode -ne [Microsoft.Azure.Test.HttpRecorder.HttpRecorderMode]::Playback) {
+        Remove-AzResourceGroup -Name $rgname -Force
+    }
+}