+++ src/OperationalInsights/OperationalInsights/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated legacy code to apply new generated SDK
+* Added parameters for `Set-AzOperationalInsightsWorkspace` and `New-AzOperationalInsightsWorkspace`
+* Created cmdlets for Linked Stoarge Account
+* Created cmdlets for Clusters and Linked Service
 
 ## Version 1.3.4
 * Update references in .psd1 to use relative path